\name{examples.data}
\alias{example1.data}
\alias{example2.data}
\alias{example3.data}
\alias{example4.data}
\alias{example6.data}
\alias{example8.data}
\alias{example1.initials}
\alias{example2.initials}
\alias{example3.initials}
\alias{example4.initials}
\alias{example6.initials}
\alias{example8.initials}
\docType{data}
\title{
  Datasets for scaRabee demo.
}
\description{
  Datasets for scaRabee demo.
}
\usage{
  data(example1.data)
  data(example1.initials)
  
  data(example2.data)
  data(example2.initials)
  
  data(example3.data)
  data(example3.initials)
  
  data(example4.data)
  data(example4.initials)
  
  data(example6.data)
  data(example6.initials)
  
  data(example8.data)
  data(example8.initials)
}
\format{
  All example#.data are data.frame of observations and dosing events organized
  as typical scaRabee datasets.
  
  All example#.initials are data.frame of parameter names and values organized
  as typical scaRabee parameter files.
}
\details{
  These datasets are intended to be used in the scaRabee package demos.
}

\examples{
data(example1.data)
data(example1.initials)
}
\keyword{datasets}
